Extract signature from crosscut

x3p file of a 3d topological surface is processed at surface crosscut y,

cc_get_signature(ccdata, span1 = 125, span2 = 5, resolution = 0.645)

Arguments

ccdata

crosscut as returned from x3p_crosscut

span1

The span for the loess fit to get from the profile to the raw signature

span2

The span for the loess fit to smooth the raw signature

Value

data frame

Examples

x3p <- x3p_read("~/Documents/CSAFE/Wirecutter/data/Aluminum Wires renamed/T1AW-LI-R1-B01.x3p")
#> Error in x3p_read("~/Documents/CSAFE/Wirecutter/data/Aluminum Wires renamed/T1AW-LI-R1-B01.x3p"): could not find function "x3p_read"
line1 <- x3p %>% x3p_extract_profile()
#> Error in x3p_extract_profile(.): could not find function "x3p_extract_profile"
line1 <- line1$line_df
#> Error in eval(expr, envir, enclos): object 'line1' not found
signature <- cc_get_signature(line1)
#> Error in group_by(., x): object 'line1' not found